requirements.txt
Flask==3.0.0
gunicorn
pytz
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
requests
email-validator  # optional: for future email form checks
